turns out to be easier to inhale all of curve1174.rs into mod.rs file, and dummy up a test routine
We will need to provide compressed point representations on curve1174 for use by client code. Internally, the curve1174 lib will
work with offset fields, multi-u64 representations of field and group elements, and a whole bunch of esoteric Bernstein-stuff.
We need to provide a more human friendly I/O format for client code.
